% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{AvailableDictionaries}
\alias{AvailableDictionaries}
\title{List available precomputed dictionaries}
\usage{
AvailableDictionaries()
}
\value{
A data frame with columns \code{name} and \code{description}
}
\description{
Returns a table of dictionary names and descriptions
}
