% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ConvertGenes}
\alias{ConvertGenes}
\title{Convert genes}
\usage{
ConvertGenes(dictionary, from = "ENSEMBL", to = "SYMBOL")
}
\arguments{
\item{dictionary}{Dictionary obtained from LearnDictionary()}

\item{from}{Key type to convert from (default 'ENSEMBL')}

\item{to}{Key type to convert to (default 'SYMBOL')}
}
\value{
Dictionary with re-named genes; genes that couldn't be converted are dropped
}
\description{
Convert genes in a dictionary object
}
\examples{
\dontrun{
dictionary <- ConvertGenes(
dictionary
)
}

}
