% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{InstallPrecomputedDictionary}
\alias{InstallPrecomputedDictionary}
\title{Install precomputed dictionary}
\usage{
InstallPrecomputedDictionary(name)
}
\arguments{
\item{name}{Name of dataset; see available datasets with AvailableData()}
}
\value{
No return value; installs requested dictionary
}
\description{
Installs requested precomputed dictionary
}
