% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{LoadPrecomputedDictionary}
\alias{LoadPrecomputedDictionary}
\title{Load pre-computed dictionary}
\usage{
LoadPrecomputedDictionary(name, gene_key = NULL)
}
\arguments{
\item{name}{Name of pre-computed dictionary}

\item{gene_key}{Optional gene key to convert to (default NULL); only human annotations supported}
}
\value{
A set of pre-computed dictionary
}
\description{
Load in pre-computed dictionary.
}
\examples{
\dontrun{
gwps_dictionary <- LoadPrecomputedDictionary(
name = 'gwps'
)

gwps_dictionary <- LoadPrecomputedDictionary(
name = 'gwps',
gene_key = 'SYMBOL'
)
}

}
